composer.json added psr-4 autoload added
